% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findsamplesize.R
\name{find_sample_size}
\alias{find_sample_size}
\title{Find Optimal Sample Size}
\usage{
find_sample_size(
  np = 0,
  ni = 2000,
  dsearch = 5,
  power_calc_fn,
  proportion_vector = 1,
  pwr,
  ...
)
}
\arguments{
\item{np}{Integer. Lower bound of the sample size search range (default: 0).}

\item{ni}{Integer. Upper bound of the sample size search range (default: 2000).}

\item{dsearch}{Numeric. Precision of the binary search. Larger values lead to finer resolution and more iterations (default: 5).}

\item{power_calc_fn}{Function. The power-calculating function within the environment that the user should already have.
The results should be in a named list with the power value listed under power "return(list(power=final_power))".}

\item{proportion_vector}{Numeric vector. Proportion of individuals allocated to each basket/arm in multi-arm trials (default: 1).}

\item{pwr}{Numeric. Nominal/expected power (e.g., 0.8 for 80% power).}
}
\value{
Numeric. The optimal sample size.
}
\description{
This function uses a binary search algorithm to find the optimal sample size required to achieve a specified power
for a statistical test. The function takes the optimal sample size from the range of sample size values provided by the binary
search helper function.
}
