% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_rec.R
\name{n_rec}
\alias{n_rec}
\title{Recursive Binary Search for Sample Size Range}
\usage{
n_rec(np, ni, power_calc_fn, dsearch = 1, pwr, p, n, ...)
}
\arguments{
\item{np}{Integer. Lower bound of the sample size range we are searching through.}

\item{ni}{Integer. Upper bound of the sample size range we are searching through.}

\item{power_calc_fn}{Function. The power-calculating function within the environment that the user should already have.
The results should be in a named list with the power value listed under power "return(list(power=final_power))".}

\item{dsearch}{Numeric. Precision of the binary search. Larger values lead to finer resolution and more iterations.}

\item{pwr}{Numeric. Nominal/expected power (e.g., 0.8 for 80\% power).}

\item{p}{Numeric. The current iterate power value calculated at the midpoint sample size.}

\item{n}{Integer. The current iterate midpoint sample size.}
}
\value{
A vector of sample sizes within the range `[np, ni]` that meet the specified precision.
}
\description{
This is a helper function that performs a binary search using recursion to find a range of sample sizes
that meet a specified power criterion. This binary search function helps narrow down the range of possible sample sizes.
}
\examples{
\dontrun{
n_rec(
  np = 10,
  ni = 1000,
  power_calc_fn = function(total_individuals, ...) {
    return(list(power = 0.8))
  },
  dsearch = 5,
  pwr = 0.8,
  p = 0.75,
  n = 500
)}
}
\author{
Desmond Zeya Chen (Contact via GitHub Issues)
}
